refractor - the viz is in its own component
